refactored BuildRequest, introduced IssueAcceptor
