+++ plugins/org.eclipse.n4js.ide/src/org/eclipse/n4js/ide/xtext/server/XLanguageServerImpl.java
@@ -10,7 +10,6 @@ package org.eclipse.n4js.ide.xtext.server;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -36,8 +35,6 @@ import org.eclipse.lsp4j.CompletionItem;
 import org.eclipse.lsp4j.CompletionList;
 import org.eclipse.lsp4j.CompletionOptions;
 import org.eclipse.lsp4j.CompletionParams;
-import org.eclipse.lsp4j.Diagnostic;
-import org.eclipse.lsp4j.DiagnosticSeverity;
 import org.eclipse.lsp4j.DidChangeConfigurationParams;
 import org.eclipse.lsp4j.DidChangeTextDocumentParams;
 import org.eclipse.lsp4j.DidChangeWatchedFilesParams;
@@ -62,9 +59,7 @@ import org.eclipse.lsp4j.InitializeResult;
 import org.eclipse.lsp4j.InitializedParams;
 import org.eclipse.lsp4j.Location;
 import org.eclipse.lsp4j.LocationLink;
-import org.eclipse.lsp4j.Position;
 import org.eclipse.lsp4j.PrepareRenameResult;
-import org.eclipse.lsp4j.PublishDiagnosticsParams;
 import org.eclipse.lsp4j.Range;
 import org.eclipse.lsp4j.ReferenceParams;
 import org.eclipse.lsp4j.RenameCapabilities;
@@ -98,11 +93,9 @@ import org.eclipse.lsp4j.services.TextDocumentService;
 import org.eclipse.lsp4j.services.WorkspaceService;
 import org.eclipse.n4js.hlc.base.HeadlessExtensionRegistrationHelper;
 import org.eclipse.n4js.ide.server.ProjectStatePersister;
-import org.eclipse.n4js.ide.validation.N4JSIssue;
 import org.eclipse.n4js.ide.xtext.server.XBuildManager.XBuildable;
 import org.eclipse.n4js.ide.xtext.server.findReferences.XWorkspaceResourceAccess;
 import org.eclipse.n4js.ide.xtext.server.rename.XIRenameService;
-import org.eclipse.xtext.diagnostics.Severity;
 import org.eclipse.xtext.findReferences.IReferenceFinder;
 import org.eclipse.xtext.ide.server.Document;
 import org.eclipse.xtext.ide.server.ICapabilitiesContributor;
@@ -136,13 +129,11 @@ import org.eclipse.xtext.resource.XtextResource;
 import org.eclipse.xtext.resource.XtextResourceSet;
 import org.eclipse.xtext.resource.impl.ResourceDescriptionsProvider;
 import org.eclipse.xtext.util.CancelIndicator;
-import org.eclipse.xtext.validation.Issue;
 import org.eclipse.xtext.workspace.IProjectConfig;
 import org.eclipse.xtext.workspace.ISourceFolder;
 import org.eclipse.xtext.xbase.lib.IterableExtensions;
 
 import com.google.common.base.Objects;
-import com.google.common.collect.ComparisonChain;
 import com.google.common.collect.FluentIterable;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
@@ -182,6 +173,9 @@ public class XLanguageServerImpl implements LanguageServer, WorkspaceService, Te
 
 	@Inject
 	private ILanguageServerShutdownAndExitHandler shutdownAndExitHandler;
+
+	@Inject
+	private IssueAcceptor issueAcceptor;
 
 	private XWorkspaceManager workspaceManager;
 
@@ -247,7 +241,7 @@ public class XLanguageServerImpl implements LanguageServer, WorkspaceService, Te
 
 		access.addBuildListener(this);
 
-		workspaceManager.initialize(baseDir, this::publishDiagnostics);
+		workspaceManager.initialize(baseDir);
 		initializeResult = new InitializeResult();
 		initializeResult.setCapabilities(createServerCapabilities(params));
 
@@ -383,6 +377,7 @@ public class XLanguageServerImpl implements LanguageServer, WorkspaceService, Te
 	@Override
 	public void connect(LanguageClient client) {
 		this.client = client;
+		issueAcceptor.connect(client);
 	}
 
 	@Override
@@ -514,98 +509,6 @@ public class XLanguageServerImpl implements LanguageServer, WorkspaceService, Te
 			workspaceManager.refreshWorkspaceConfig(CancelIndicator.NullImpl);
 			return null;
 		}, (a, b) -> null);
-	}
-
-	private void publishDiagnostics(URI uri, Iterable<? extends Issue> issues) {
-		PublishDiagnosticsParams publishDiagnosticsParams = new PublishDiagnosticsParams();
-		publishDiagnosticsParams.setUri(uriExtensions.toUriString(uri));
-		List<Diagnostic> diags = toDiagnostics(uri, issues);
-		publishDiagnosticsParams.setDiagnostics(diags);
-		client.publishDiagnostics(publishDiagnosticsParams);
-	}
-
-	/**
-	 * Convert the given issues to diagnostics. Does not return issues in files that are neither in the workspace nor
-	 * currently opened in the editor. Does not return any issue with severity {@link Severity#IGNORE ignore}.
-	 */
-	protected List<Diagnostic> toDiagnostics(URI uri, Iterable<? extends Issue> issues) {
-		if (!workspaceManager.isDocumentOpen(uri)) {
-			// Closed documents need to exist in the current workspace
-			IProjectConfig projectConfig = workspaceManager.getWorkspaceConfig().findProjectContaining(uri);
-			ISourceFolder sourceFolder = projectConfig.findSourceFolderContaining(uri);
-			if (sourceFolder == null) {
-				return Collections.emptyList();
-			}
-		}
-
-		List<Diagnostic> sortedDiags = new ArrayList<>();
-		for (Issue issue : issues) {
-			if (issue.getSeverity() != Severity.IGNORE) {
-				sortedDiags.add(toDiagnostic(issue));
-			}
-		}
-
-		// Sort issues according to line and position
-		final Comparator<Diagnostic> comparator = new Comparator<>() {
-			@Override
-			public int compare(Diagnostic d1, Diagnostic d2) {
-				Position p1 = d1.getRange().getStart();
-				Position p2 = d2.getRange().getStart();
-				int result = ComparisonChain.start()
-						.compare(p1.getLine(), p2.getLine())
-						.compare(p2.getCharacter(), p2.getCharacter())
-						.result();
-				return result;
-			}
-		};
-
-		Collections.sort(sortedDiags, comparator);
-
-		return sortedDiags;
-	}
-
-	/**
-	 * Convert the given issue to a diagnostic.
-	 */
-	protected Diagnostic toDiagnostic(Issue issue) {
-		Diagnostic result = new Diagnostic();
-		result.setCode(issue.getCode());
-		result.setMessage(issue.getMessage());
-		result.setSeverity(toDiagnosticSeverity(issue.getSeverity()));
-
-		Position start = new Position(issue.getLineNumber(), issue.getColumn());
-		Position end = null;
-		if (issue instanceof N4JSIssue) {
-			N4JSIssue n4jsIssue = (N4JSIssue) issue;
-			end = new Position(n4jsIssue.getLineNumberEnd(), n4jsIssue.getColumnEnd());
-		} else {
-			URI uri = issue.getUriToProblem();
-			Document doc = workspaceManager.getDocument(uri);
-			doc.getPosition(issue.getOffset() + issue.getLength());
-		}
-
-		result.setRange(new Range(start, end));
-		return result;
-	}
-
-	/**
-	 * Convert the severity to a lsp {@link DiagnosticSeverity}.
-	 *
-	 * Defaults to severity {@link DiagnosticSeverity#Hint hint}.
-	 */
-	protected DiagnosticSeverity toDiagnosticSeverity(Severity severity) {
-		switch (severity) {
-		case ERROR:
-			return DiagnosticSeverity.Error;
-		case IGNORE:
-			return DiagnosticSeverity.Hint;
-		case INFO:
-			return DiagnosticSeverity.Information;
-		case WARNING:
-			return DiagnosticSeverity.Warning;
-		default:
-			return DiagnosticSeverity.Hint;
-		}
 	}
 
 	@Override
